Fixed #236
Refactored some classes to make the minimum impact oer what we already have done so far.  Created two new tests to assert it is working fine, but all the business logic previous made is almost untouched, so I think it would be a waste of resource to replicate many tests that we already have for without using * (any) and using it, but we can do it.