Correctly count semantically equal objects but with different pointers
Found that WeakHashMap still uses object equality instead of reference
equality, which is of course intentional but not the semantics we need
for approximating memory usage which relies on references of course.
Additionally, many tests were designed with assumption that same
reference would be counted many times, so these had to be updated to
create copies of objects as would occur with real queries.